Multiplying weights by 0.01 improves accuracy. This could be due to the loss function. Also adding functionality for float32. Added printing of training acc

+++ fault_tolerant_ml/ml/linear_model/logistic.py
@@ -42,7 +42,7 @@ class LogisticRegression(BaseEstimator, LinearClassifierMixin):
         n_classes = len(self.classes_)
 
         # Initialize parameters
-        self.theta = np.random.randn(n_features, n_classes) * 0.01
+        self.theta = (np.random.randn(n_features, n_classes) * 0.01).astype(np.float32)
         
         i = 0
         delta = 1.0
